+++ wayland-server/src/lib.rs
@@ -293,6 +293,19 @@ impl<I: Resource> Weak<I> {
         I::from_id(&d_handle, self.id.clone())
     }
 
+    /// Check if this resource is still alive
+    ///
+    /// This will return `false` if either:
+    /// - the object represented by this handle has already been destroyed at the protocol level
+    /// - the Wayland connection has already been closed
+    #[inline]
+    pub fn is_alive(&self) -> bool {
+        let Some(handle) = self.handle.upgrade() else {
+            return false;
+        };
+        handle.object_info(self.id.clone()).is_ok()
+    }
+
     /// The underlying [`ObjectId`]
     pub fn id(&self) -> ObjectId {
         self.id.clone()